refactor: move extract enums to separate file
